Tehty tehtävään 4.22 asti

+++ osa4/blogilista/controllers/blogs.js
@@ -1,14 +1,19 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', {username: 1, name: 1 })
+
   response.json(blogs)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
+  
   if (blog) {
     response.json(blog)
   } else {
@@ -16,19 +21,18 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
-
-  const user = await User.findById(body.userId)
-
+  
+  const user = request.user
   if (!user) {
-    return response.status(400).json({ error: 'userId missing or invalid' })
+    return response.status(401).json({ error: 'token invalid or missing' })
   }
 
   const blog = new Blog({
+    url: body.url,
     title: body.title,
     author: body.author,
-    url: body.url,
     likes: body.likes || 0,
     user: user._id
   })
@@ -40,18 +44,32 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'token invalid or missing' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'only the creator can delete a blog' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
 
   const blog = {
+    url: body.url,
     title: body.title,
     author: body.author,
-    url: body.url,
     likes: body.likes
   }
